refactor: separate presentation from logic with container/view pattern
- Implement Container/Presentational pattern with ESLint enforcement
- Extract useLocalStorage hook for DRY code reusability
- Fix duplicate timer bug in NotificationProvider
- Move auto-focus logic from view to container (zero lint warnings)
- Fix useEffect dependencies in ShellPromptView
- Update AGENTS.md with comprehensive architecture documentation
- Add localStorage clearing in tests to prevent interference

Architecture improvements:
- Containers (src/containers/): Manage state, side effects, business logic
- Views (src/components/*/View.tsx): Pure presentational components
- ESLint enforces pattern: no useState in src/components/
- All side effects (focus, localStorage, timers) now in containers

Code quality:
- All 70 tests passing with 99.24% coverage
- Zero lint warnings or errors
- Production build successful
- Perfect Container/Presentational separation
