+++ src/hooks/useEvent.test.ts
@@ -1,5 +1,5 @@
 import { act, renderHook } from '@testing-library/react';
-import type { AchievementUnlocked } from '../components/Achievements/Achievements';
+import type { AchievementUnlocked } from '../containers/AchievementProvider';
 import type { ConsoleLine } from '../components/ConsoleOutput/ConsoleOutput';
 import { useEvent } from './useEvent';
