+++ src/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import { NotificationProvider } from './components/Notification/Notification';
-import { AchievementProvider } from './components/Achievements/Achievements';
+import { NotificationProvider } from './containers/NotificationProvider';
+import { AchievementProvider } from './containers/AchievementProvider';
 
 const root = ReactDOM.createRoot(
     document.getElementById('root') as HTMLElement
